+++ core/blockchain/nullchain/server.go
@@ -28,9 +28,14 @@ import (
 
 var ErrInvalidRequest = errors.New("invalid request")
 
+const (
+	NullChainStatusReady     = "chain-ready"
+	NullChainStatusReplaying = "chain-replaying"
+)
+
 func (n *NullBlockchain) Stop() error {
-	if r, ok := n.app.(*Replayer); ok {
-		if err := r.Stop(); err != nil {
+	if n.replayer != nil {
+		if err := n.replayer.Stop(); err != nil {
 			n.log.Error("failed to stop nullchain replayer")
 		}
 	}
@@ -53,16 +58,18 @@ func (n *NullBlockchain) Start() error {
 }
 
 func (n *NullBlockchain) StartServer() error {
+	n.srv = &http.Server{Addr: net.JoinHostPort(n.cfg.IP, strconv.Itoa(n.cfg.Port))}
+	http.HandleFunc("/api/v1/forwardtime", n.handleForwardTime)
+	http.HandleFunc("/api/v1/status", n.status)
+
+	n.log.Info("starting time-forwarding server", logging.String("addr", n.srv.Addr))
+	go n.srv.ListenAndServe()
+
 	n.log.Info("starting blockchain")
 	if err := n.StartChain(); err != nil {
 		return err
 	}
 
-	n.srv = &http.Server{Addr: net.JoinHostPort(n.cfg.IP, strconv.Itoa(n.cfg.Port))}
-	http.HandleFunc("/api/v1/forwardtime", n.handleForwardTime)
-
-	n.log.Info("starting time-forwarding server", logging.String("addr", n.srv.Addr))
-	go n.srv.ListenAndServe()
 	return nil
 }
 
@@ -110,6 +117,45 @@ func (n *NullBlockchain) handleForwardTime(w http.ResponseWriter, r *http.Reques
 		http.Error(w, err.Error(), http.StatusBadRequest)
 	}
 
-	// Do the dance
-	n.ForwardTime(d)
+	if n.replaying.Load() {
+		http.Error(w, ErrChainReplaying.Error(), http.StatusServiceUnavailable)
+	}
+
+	// we need to call ForwardTime in a different routine so that if it panics it stops the node instead of just being
+	// caught in the http-handler recover. But awkwardly it seems like the vega-sim relies on the http-request not
+	// returning until the time-forward has finished, so we need to preserve that for now.
+	done := make(chan struct{})
+	go func() {
+		n.ForwardTime(d)
+		done <- struct{}{}
+	}()
+	<-done
+}
+
+// status returns the status of the nullchain, whether it is replaying or whether its ready to go.
+func (n *NullBlockchain) status(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+
+	resp := struct {
+		Status string `json:"status"`
+	}{
+		Status: NullChainStatusReady,
+	}
+
+	if n.replaying.Load() {
+		resp.Status = NullChainStatusReplaying
+	}
+
+	buf, err := json.Marshal(resp)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	_, err = w.Write(buf)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 }